remove persistent topnav and add a standard rightnav ctrl
