+++ ui/layout.jsx
@@ -7,8 +7,6 @@ import ModalFlow from './com/modals/flow'
 import Welcome from './com/forms/welcome'
 import ProfileName from './com/forms/profile-name'
 import ProfileImage from './com/forms/profile-image'
-import Issues from './com/issues'
-import SearchPalette from './com/search-palette'
 import FindBar from './com/findbar'
 
 const SETUP_FORMS = [Welcome, ProfileName, ProfileImage]
@@ -21,9 +19,6 @@ export default class Layout extends React.Component {
     // listen for app change-events that should update our state
     const refresh = () => { this.setState(this.buildState()) }
     app.on('update:all', refresh)
-    app.on('update:indexCounts', refresh)
-    app.on('update:isWifiMode', refresh)
-    app.on('focus:search', this.focusSearch.bind(this))
     app.on('focus:find', this.focusFind.bind(this))
     app.on('modal:setup', isOpen => { this.setState({ setupIsOpen: isOpen }); app.fetchLatestState() })
     app.on('find:next', this.doFind.bind(this, true))
@@ -36,17 +31,9 @@ export default class Layout extends React.Component {
   buildState() {
     // copy over app state
     return {
-      isWifiMode: app.isWifiMode,
-      indexCounts: app.indexCounts||{},
-      user: app.user,
-      users: app.users,
       setupIsOpen: app.user.needsSetup,
       isComposerOpen: app.isComposerOpen
     }
-  }
-
-  focusSearch() {
-     this.refs.search.focus()
   }
 
   focusFind() {
@@ -66,34 +53,8 @@ export default class Layout extends React.Component {
   }
 
   render() {
-    const isElectron = !!window.electron
-    const location = this.props.location.pathname
-    const isWifiMode = this.state.isWifiMode
-
-    const NavLink = (props) => {
-      const selected = props.selected || (props.to === location)
-      const cls = classNames(props.className||'', 'ctrl', { selected }, props.hint ? ('hint--'+props.hint) : '')
-      const count = props.count ? <div className="count">{props.count}</div> : ''
-      return <Link className={cls} to={props.to} data-hint={props.title}>
-        <i className={'fa fa-'+props.icon} />
-        <span className="label">{props.label}</span> {count}
-      </Link>
-    }
-
     return <div className="layout-rows">
       <ModalFlow className="fullheight" Forms={SETUP_FORMS} isOpen={this.state.setupIsOpen} onClose={this.onSetupClose.bind(this)} />
-      <div className="toolbar titlebar flex">
-        <div>
-          { isElectron
-            ? <a className="ctrl back" onClick={this.onClickBack}><i className="fa fa-angle-left" /></a>
-            : '' }
-          <NavLink className="home" to="/" icon="home" title="Home" hint="bottom-right" />
-        </div>
-        <div className="flex-fill"><SearchPalette ref="search"/></div>
-        <NavLink to="/digs" icon="hand-peace-o" count={app.indexCounts.digsUnread} title="Digs on your posts" hint="bottom" />
-        <NavLink to="/sync" icon="cloud-download" title="Network sync status" hint="bottom" />
-        <NavLink to="/data" icon="database" title="Raw database feed" hint="bottom-left" />
-      </div>
       <div className="layout-columns">
         <div id="mainview">{this.props.children}</div>
       </div>
