+++ ui/less/com/rightnav.less
@@ -4,16 +4,14 @@
   right: 0;
   z-index: 1;
 
-  a {
-    .elevation-1;
+  a.btn {
     display: block;
-    background: #fafafa;
     margin: 10px 10px 0 0;
     padding: 10px;
-    color: gray;
-    &:hover {
-      background: #fff;
-      text-decoration: none;
-    }
+  }
+  .toolbar {
+    margin: 10px 10px 0 0;
+    background: #fdfdfd;
+    border: 0;
   }
 }
