+++ ui/views/follows.jsx
@@ -2,6 +2,7 @@
 import React from 'react'
 import { Link } from 'react-router'
 import LeftNav from '../com/leftnav'
+import RightNav from '../com/rightnav'
 import MsgList from '../com/msg-list'
 import Oneline from '../com/msg-view/oneline'
 import app from '../lib/app'
@@ -22,10 +23,10 @@ export default class Follows extends React.Component {
   }
 
   render() {
-    const RightNav = props => {
-      return <div className="rightnav">
-        <a onClick={this.onMarkAllRead.bind(this)} href="javascript:"><i className="fa fa-envelope" /> Mark all read</a>
-      </div>
+    const ThisRightNav = props => {
+      return <RightNav>
+        <a className="btn" onClick={this.onMarkAllRead.bind(this)} href="javascript:"><i className="fa fa-envelope" /> Mark all read</a>
+      </RightNav>
     }
 
     return <div id="follows">
@@ -36,7 +37,7 @@ export default class Follows extends React.Component {
         composer composerProps={{ isPublic: true }}
         ListItem={Oneline} listItemProps={{ userPic: true }}
         LeftNav={LeftNav} leftNavProps={{ location: this.props.location }}
-        RightNav={RightNav}
+        RightNav={ThisRightNav}
         live={{ gt: [Date.now(), null] }}
         emptyMsg="Nobody has followed you yet."
         source={app.ssb.patchwork.createFollowStream}
