+++ ui/views/private.jsx
@@ -1,6 +1,7 @@
 'use babel'
 import React from 'react'
 import LeftNav from '../com/leftnav'
+import RightNav from '../com/rightnav'
 import MsgList from '../com/msg-list'
 import Oneline from '../com/msg-view/oneline'
 import app from '../lib/app'
@@ -21,10 +22,10 @@ export default class PrivatePosts extends React.Component {
   }
 
   render() {
-    const RightNav = props => {
-      return <div className="rightnav">
-        <a onClick={this.onMarkAllRead.bind(this)} href="javascript:"><i className="fa fa-envelope" /> Mark all read</a>
-      </div>
+    const ThisRightNav = props => {
+      return <RightNav>
+        <a className="btn" onClick={this.onMarkAllRead.bind(this)} href="javascript:"><i className="fa fa-envelope" /> Mark all read</a>
+      </RightNav>
     }
 
     return <div id="private">
@@ -35,7 +36,7 @@ export default class PrivatePosts extends React.Component {
         composer composerProps={{ isPublic: false }}
         ListItem={Oneline} listItemProps={{ userPic: true }}
         LeftNav={LeftNav} leftNavProps={{location: this.props.location}}
-        RightNav={RightNav}
+        RightNav={ThisRightNav}
         live={{ gt: [Date.now(), null] }}
         emptyMsg="You have no private messages."
         source={app.ssb.patchwork.createPrivatePostStream}
