+++ ui/views/public.jsx
@@ -1,6 +1,7 @@
 'use babel'
 import React from 'react'
 import LeftNav from '../com/leftnav'
+import RightNav from '../com/rightnav'
 import MsgList from '../com/msg-list'
 import Oneline from '../com/msg-view/oneline'
 import app from '../lib/app'
@@ -42,6 +43,7 @@ export default class PublicPosts extends React.Component {
         openMsgEvent
         composer composerProps={{ isPublic: true, channel: channel }}
         LeftNav={LeftNav} leftNavProps={{location: this.props.location}}
+        RightNav={RightNav}
         ListItem={Oneline} listItemProps={{ userPic: true }}
         live={{ gt: [Date.now(), null] }}
         emptyMsg={(channel) ? ('No posts on "'+channel+'"... yet!') : 'Your feed is empty.'}
